requirements.txt
flask
pytz
mysql-connector-python